Refactor validate_default_by_typ into defaults::validator
