+++ components/support/nimbus-fml/src/intermediate_representation.rs
@@ -1,9 +1,9 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-use crate::defaults::DefaultsMerger;
+use crate::defaults::{DefaultsMerger, DefaultsValidator};
 use crate::error::FMLError::InvalidFeatureError;
-use crate::error::{did_you_mean, FMLError, Result};
+use crate::error::{FMLError, Result};
 use crate::frontend::{AboutBlock, FeatureMetadata};
 use crate::util::loaders::FilePath;
 use anyhow::{bail, Error, Result as AnyhowResult};
@@ -117,6 +117,13 @@ impl TypeRef {
             // This may be done by adding a branch to this match and adding a `preference_getter` to
             // the `OptionalCodeType`.
             _ => false,
+        }
+    }
+
+    pub(crate) fn name(&self) -> Option<&str> {
+        match self {
+            Self::Enum(s) | Self::Object(s) => Some(s),
+            _ => None,
         }
     }
 }
@@ -450,15 +457,13 @@ impl FeatureManifest {
     }
 
     fn validate_defaults(&self) -> Result<()> {
+        let validator = DefaultsValidator::new(&self.enum_defs, &self.obj_defs);
         for object in self.iter_object_defs() {
-            for prop in &object.props {
-                let path = format!("objects/{}.{}", object.name, prop.name);
-                self.validate_prop_defaults(&path, prop)?;
-            }
+            validator.validate_object_def(object)?;
         }
         for feature in self.iter_feature_defs() {
             self.validate_feature_structure(feature)?;
-            self.validate_feature_def(feature)?;
+            validator.validate_feature_def(feature)?;
         }
         Ok(())
     }
@@ -474,196 +479,6 @@ impl FeatureManifest {
             }
         }
         Ok(())
-    }
-
-    fn validate_feature_def(&self, feature_def: &FeatureDef) -> Result<()> {
-        for prop in &feature_def.props {
-            let path = format!("features/{}.{}", feature_def.name, prop.name);
-            let literals = vec![prop.name.to_string()];
-            self.validate_default_by_typ(&path, &literals, &prop.typ, &prop.default)?;
-        }
-        Ok(())
-    }
-
-    fn validate_prop_defaults(&self, path: &str, prop: &PropDef) -> Result<()> {
-        let literals = Default::default();
-        self.validate_default_by_typ(path, &literals, &prop.typ, &prop.default)
-    }
-
-    pub fn validate_default_by_typ(
-        &self,
-        path: &str,
-        literals: &Vec<String>,
-        type_ref: &TypeRef,
-        default: &Value,
-    ) -> Result<()> {
-        let add_literals = |new: &[String]| -> Vec<String> {
-            let mut clone = literals.clone();
-            for s in new {
-                clone.push(s.clone());
-            }
-            clone
-        };
-        let add_literal = |new: String| -> Vec<String> {
-            let mut clone = literals.clone();
-            clone.push(new);
-            clone
-        };
-        match (type_ref, default) {
-            (TypeRef::Boolean, Value::Bool(_))
-            | (TypeRef::BundleImage(_), Value::String(_))
-            | (TypeRef::BundleText(_), Value::String(_))
-            | (TypeRef::String, Value::String(_))
-            | (TypeRef::Int, Value::Number(_))
-            | (TypeRef::Option(_), Value::Null) => Ok(()),
-            (TypeRef::Option(inner), v) => {
-                if let TypeRef::Option(_) = inner.as_ref() {
-                    return Err(FMLError::ValidationError(
-                        path.to_string(),
-                        "Nested options".into(),
-                    ));
-                }
-                self.validate_default_by_typ(path, literals, inner, v)
-            }
-            (TypeRef::Enum(enum_name), Value::String(s)) => {
-                let enum_def = self.find_enum(enum_name).ok_or_else(|| {
-                    FMLError::ValidationError(
-                        path.to_string(),
-                        format!("Type `{enum_name}` is not a type. Perhaps you need to declare an enum of that name.")
-                    )
-                })?;
-                let mut valid = HashSet::new();
-                for variant in enum_def.variants() {
-                    let name = variant.name();
-                    if *s == name {
-                        return Ok(());
-                    }
-                    valid.insert(name);
-                }
-                Err(FMLError::FeatureValidationError {
-                    path: path.to_string(),
-                    message: format!("\"{s}\" is not a valid {enum_name}{}", did_you_mean(valid)),
-                    literals: add_literal(format!("\"{s}\"")),
-                })
-            }
-            (TypeRef::EnumMap(enum_type, map_type), Value::Object(map)) => {
-                let enum_name = if let TypeRef::Enum(name) = enum_type.as_ref() {
-                    name.clone()
-                } else {
-                    unreachable!()
-                };
-                // We first validate that the keys of the map cover all all the enum variants, and no more or less
-                let enum_def = self.find_enum(&enum_name).ok_or_else(|| {
-                    FMLError::ValidationError(
-                        path.to_string(),
-                        format!("Type `{}` is not a type. Perhaps you need to declare an enum of that name.", enum_name)
-                    )
-                })?;
-                let mut seen = HashSet::new();
-                let mut unseen = HashSet::new();
-                let mut valid = HashSet::new();
-                for variant in enum_def.variants() {
-                    let nm = variant.name();
-                    valid.insert(nm.clone());
-
-                    let map_value = map.get(&nm);
-                    match (map_type.as_ref(), map_value) {
-                        (TypeRef::Option(_), None) => (),
-                        (_, None) => {
-                            unseen.insert(variant.name());
-                        }
-                        (_, Some(inner)) => {
-                            let path = format!("{path}[{}#{nm}]", enum_def.name);
-                            let literals = add_literals(&["{".to_string(), format!("\"{nm}\"")]);
-                            self.validate_default_by_typ(&path, &literals, map_type, inner)?;
-                            seen.insert(nm);
-                        }
-                    }
-                }
-
-                if !unseen.is_empty() {
-                    return Err(FMLError::FeatureValidationError {
-                        path: path.to_string(),
-                        message: format!("Enum map {enum_name} is missing values for {unseen:?}"),
-                        // Can we be more specific that just the opening brace?
-                        literals: add_literal("{".to_string()),
-                    });
-                }
-                for map_key in map.keys() {
-                    if !seen.contains(map_key) {
-                        return Err(FMLError::FeatureValidationError {
-                            path: path.to_string(),
-                            message: format!("Invalid key \"{map_key}\"{}", did_you_mean(valid)),
-                            literals: add_literals(&["{".to_string(), format!("\"{map_key}\"")]),
-                        });
-                    }
-                }
-                Ok(())
-            }
-            (TypeRef::StringMap(map_type), Value::Object(map)) => {
-                for (key, value) in map {
-                    let path = format!("{path}['{key}']");
-                    let literals = add_literals(&["{".to_string(), format!("\"{key}\"")]);
-                    self.validate_default_by_typ(&path, &literals, map_type, value)?;
-                }
-                Ok(())
-            }
-            (TypeRef::List(list_type), Value::Array(arr)) => {
-                let mut literals = add_literal("[".to_string());
-                for (index, value) in arr.iter().enumerate() {
-                    let path = format!("{path}['{index}']");
-                    self.validate_default_by_typ(&path, &literals, list_type, value)?;
-                    literals.push(",".to_string());
-                }
-                Ok(())
-            }
-            (TypeRef::Object(obj_name), Value::Object(map)) => {
-                let obj_def = self.find_object(obj_name).ok_or_else(|| {
-                    FMLError::ValidationError(
-                        path.to_string(),
-                        format!("Object {obj_name} is not defined in the manifest"),
-                    )
-                })?;
-                let mut valid = HashSet::new();
-                let mut unseen = HashSet::new();
-                let path = format!("{path}#{obj_name}");
-                for prop in &obj_def.props {
-                    // We only check the defaults overriding the property defaults
-                    // from the object's own property defaults.
-                    // We check the object property defaults previously.
-                    let nm = prop.name();
-                    if let Some(map_val) = map.get(&nm) {
-                        let path = format!("{path}.{}", prop.name);
-                        let literals =
-                            add_literals(&["{".to_string(), format!("\"{}\"", &prop.name)]);
-                        self.validate_default_by_typ(&path, &literals, &prop.typ, map_val)?;
-                    } else {
-                        unseen.insert(nm.clone());
-                    }
-
-                    valid.insert(nm);
-                }
-                for map_key in map.keys() {
-                    if !valid.contains(map_key) {
-                        return Err(FMLError::FeatureValidationError {
-                            path,
-                            message: format!(
-                                "Invalid key \"{map_key}\" for object {obj_name}{}",
-                                did_you_mean(valid)
-                            ),
-                            literals: add_literal(format!("\"{map_key}\"")),
-                        });
-                    }
-                }
-
-                Ok(())
-            }
-            _ => Err(FMLError::FeatureValidationError {
-                path: path.to_string(),
-                message: format!("Mismatch between type {type_ref:?} and default {default}"),
-                literals: add_literal(default.to_string()),
-            }),
-        }
     }
 
     pub fn iter_enum_defs(&self) -> impl Iterator<Item = &EnumDef> {
@@ -780,7 +595,10 @@ impl FeatureManifest {
 
         let mut feature_def = feature_def.clone();
         merger.merge_feature_defaults(&mut feature_def, &Some(vec![feature_value.into()]))?;
-        manifest.validate_feature_def(&feature_def)?;
+
+        let validator = DefaultsValidator::new(&manifest.enum_defs, &manifest.obj_defs);
+        validator.validate_feature_def(&feature_def)?;
+
         Ok(feature_def)
     }
 }
@@ -1359,441 +1177,6 @@ pub mod unit_tests {
         }
 
         res
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_string() -> Result<()> {
-        let mut prop = PropDef::new("key", TypeRef::String, json!("default!"));
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_string.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!(100);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out, default is number when it should be string");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_int() -> Result<()> {
-        let mut prop = PropDef::new("key", TypeRef::Int, json!(100));
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_int.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!("100");
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out, default is string when it should be number");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_bool() -> Result<()> {
-        let mut prop = PropDef::new("key", TypeRef::Boolean, json!(true));
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_bool.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!("100");
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out, default is string when it should be a boolean");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_bundle_image() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::BundleImage("Icon".into()),
-            json!("IconBlue"),
-        );
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_bundle_image.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!(100);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err(
-            "Should error out, default is number when it should be a string (bundleImage string)",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_bundle_text() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::BundleText("Text".into()),
-            json!("BundledText"),
-        );
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_bundle_text.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!(100);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err(
-            "Should error out, default is number when it should be a string (bundleText string)",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_option_null() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::Option(Box::new(TypeRef::Boolean)),
-            json!(null),
-        );
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_option_null.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!(100);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err(
-            "Should error out, default is number when it should be a boolean (Optional boolean)",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_nested_options() -> Result<()> {
-        let prop = PropDef::new(
-            "key",
-            TypeRef::Option(Box::new(TypeRef::Option(Box::new(TypeRef::Boolean)))),
-            json!(true),
-        );
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_nested_options.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out since we have a nested option");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_option_non_null() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::Option(Box::new(TypeRef::Boolean)),
-            json!(true),
-        );
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        let path = format!("test_validate_prop_defaults_option_non_null.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!(100);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err(
-            "Should error out, default is number when it should be a boolean (Optional boolean)",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_enum() -> Result<()> {
-        let mut prop = PropDef::new("key", TypeRef::Enum("ButtonColor".into()), json!("blue"));
-        let enum_defs = vec![EnumDef {
-            name: "ButtonColor".into(),
-            variants: vec![
-                VariantDef {
-                    name: "blue".into(),
-                    ..Default::default()
-                },
-                VariantDef {
-                    name: "green".into(),
-                    ..Default::default()
-                },
-            ],
-            ..Default::default()
-        }];
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs.clone(), &prop);
-        let path = format!("test_validate_prop_defaults_enum.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!("green");
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!("not a valid color");
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out since default is not a valid enum variant");
-        prop.default = json!("blue");
-        prop.typ = TypeRef::Enum("DoesntExist".into());
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs, &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out since the enum definition doesn't exist for the TypeRef");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_enum_map() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::EnumMap(
-                Box::new(TypeRef::Enum("ButtonColor".into())),
-                Box::new(TypeRef::Int),
-            ),
-            json!({
-                "blue": 1,
-                "green": 22,
-            }),
-        );
-        let enum_defs = vec![EnumDef {
-            name: "ButtonColor".into(),
-            variants: vec![
-                VariantDef {
-                    name: "blue".into(),
-                    ..Default::default()
-                },
-                VariantDef {
-                    name: "green".into(),
-                    ..Default::default()
-                },
-            ],
-            ..Default::default()
-        }];
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs.clone(), &prop);
-        let path = format!("test_validate_prop_defaults_enum_map.{}", prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!({
-            "blue": 1,
-        });
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out because the enum map is missing the green key");
-        prop.default = json!({
-            "blue": 1,
-            "green": 22,
-            "red": 3,
-        });
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs, &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err("Should error out because the default includes an extra key that is not a variant of the enum (red)");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_string_map() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::StringMap(Box::new(TypeRef::Int)),
-            json!({
-                "blue": 1,
-                "green": 22,
-            }),
-        );
-        let path = format!("test_validate_prop_defaults_string_map.{}", &prop.name);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!({
-            "blue": 1,
-        });
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!({
-            "blue": 1,
-            "green": 22,
-            "red": 3,
-            "white": "AHA not a number"
-        });
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err("Should error out because the string map includes a value that is not an int as defined by the TypeRef");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_list() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::List(Box::new(TypeRef::Int)),
-            json!([1, 3, 100]),
-        );
-        let path = format!("test_validate_prop_defaults_list.{}", &prop.name);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!([1, 2, "oops"]);
-        let fm = get_one_prop_feature_manifest(vec![], vec![], &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out because one of the values in the array is not an int");
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_object() -> Result<()> {
-        let mut prop = PropDef::new(
-            "key",
-            TypeRef::Object("SampleObj".into()),
-            json!({
-                "int": 1,
-                "string": "bobo",
-                "enum": "green",
-                "list": [true, false, true],
-                "nestedObj": {
-                    "enumMap": {
-                        "blue": 1,
-                        "green": 2,
-                    }
-                },
-                "optional": 2,
-            }),
-        );
-        let obj_defs = vec![
-            ObjectDef {
-                name: "SampleObj".into(),
-                props: vec![
-                    PropDef::new("int", TypeRef::Int, json!(1)),
-                    PropDef::new("string", TypeRef::String, json!("a string")),
-                    PropDef::new("enum", TypeRef::Enum("ButtonColor".into()), json!("blue")),
-                    PropDef::new(
-                        "list",
-                        TypeRef::List(Box::new(TypeRef::Boolean)),
-                        json!([true, false]),
-                    ),
-                    PropDef::new(
-                        "optional",
-                        TypeRef::Option(Box::new(TypeRef::Int)),
-                        json!(null),
-                    ),
-                    PropDef::new(
-                        "nestedObj",
-                        TypeRef::Object("NestedObject".into()),
-                        json!({
-                            "enumMap": {
-                                "blue": 1,
-                            },
-                        }),
-                    ),
-                ],
-                ..Default::default()
-            },
-            ObjectDef {
-                name: "NestedObject".into(),
-                props: vec![PropDef::new(
-                    "enumMap",
-                    TypeRef::EnumMap(
-                        Box::new(TypeRef::Enum("ButtonColor".into())),
-                        Box::new(TypeRef::Int),
-                    ),
-                    json!({
-                        "blue": 4,
-                        "green": 2,
-                    }),
-                )],
-                ..Default::default()
-            },
-        ];
-        let enum_defs = vec![EnumDef {
-            name: "ButtonColor".into(),
-            variants: vec![
-                VariantDef {
-                    name: "blue".into(),
-                    ..Default::default()
-                },
-                VariantDef {
-                    name: "green".into(),
-                    ..Default::default()
-                },
-            ],
-            ..Default::default()
-        }];
-        let path = format!("test_validate_prop_defaults_object.{}", &prop.name);
-        let fm = get_one_prop_feature_manifest(obj_defs.clone(), enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop)?;
-        prop.default = json!({
-            "int": 1,
-            "string": "bobo",
-            "enum": "green",
-            "list": [true, false, true],
-            "nestedObj": {
-                "enumMap": {
-                    "blue": 1,
-                    "green": "Wrong type!"
-                }
-            }
-        });
-        let fm = get_one_prop_feature_manifest(obj_defs.clone(), enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop).expect_err(
-            "Should error out because the nested object has an enumMap with the wrong type",
-        );
-        prop.default = json!({
-            "int": 1,
-            "string": "bobo",
-            "enum": "green",
-            "list": [true, false, true],
-            "nestedObj": {
-                "enumMap": {
-                    "blue": 1,
-                    "green": 2,
-                }
-            },
-            "optional": 3,
-            "extra-property": 2
-        });
-        let fm = get_one_prop_feature_manifest(obj_defs.clone(), enum_defs.clone(), &prop);
-        fm.validate_prop_defaults(&path, &prop)
-            .expect_err("Should error out because the object has an extra property");
-
-        // This test is missing a `list` property. But that's ok, because we'll get it from the object definition.
-        prop.default = json!({
-            "int": 1,
-            "string": "bobo",
-            "enum": "green",
-            "nestedObj": {
-                "enumMap": {
-                    "blue": 1,
-                    "green": 2,
-                }
-            },
-            "optional": 2,
-        });
-        let fm = get_one_prop_feature_manifest(obj_defs.clone(), enum_defs.clone(), &prop);
-
-        fm.validate_prop_defaults(&path, &prop)?;
-
-        prop.default = json!({
-            "int": 1,
-            "string": "bobo",
-            "enum": "green",
-            "list": [true, false, true],
-            "nestedObj": {
-                "enumMap": {
-                    "blue": 1,
-                    "green": 2,
-                }
-            },
-        });
-        let fm = get_one_prop_feature_manifest(obj_defs, enum_defs, &prop);
-        // OK, because we are missing `optional` which is optional anyways
-        fm.validate_prop_defaults(&path, &prop)?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_validate_prop_defaults_enum_map_optional() -> Result<()> {
-        let prop = PropDef::new(
-            "key",
-            TypeRef::EnumMap(
-                Box::new(TypeRef::Enum("ButtonColor".into())),
-                Box::new(TypeRef::Option(Box::new(TypeRef::Int))),
-            ),
-            json!({
-                "blue": 1,
-            }),
-        );
-        let enum_defs = vec![EnumDef {
-            name: "ButtonColor".into(),
-            variants: vec![
-                VariantDef {
-                    name: "blue".into(),
-                    ..Default::default()
-                },
-                VariantDef {
-                    name: "green".into(),
-                    ..Default::default()
-                },
-            ],
-            ..Default::default()
-        }];
-        let fm = get_one_prop_feature_manifest(vec![], enum_defs, &prop);
-        // OK because the value is optional, and thus it's okay if it's missing (green is missing from the default)
-        let path = format!("test.{}", &prop.name);
-        fm.validate_prop_defaults(&path, &prop)?;
-        Ok(())
     }
 
     #[test]
